update dependencies, fix legacy issue

+++ run_visualise.py
@@ -69,7 +69,7 @@ for fidx, fname in enumerate(flist):
 
 	for face in faces[fidx]:
 
-		clr = max(min(face['conf']*25,255),0)
+		clr = int(max(min(face['conf']*25,255),0))
 
 		cv2.rectangle(image,(int(face['x']-face['s']),int(face['y']-face['s'])),(int(face['x']+face['s']),int(face['y']+face['s'])),(0,clr,255-clr),3)
 		cv2.putText(image,'Track %d, Conf %.3f'%(face['track'],face['conf']), (int(face['x']-face['s']),int(face['y']-face['s'])),cv2.FONT_HERSHEY_SIMPLEX,0.5,(255,255,255),2)
